Project refactor, introduced views to some queries
