+++ Library_Database_Project.egg-info/requires.txt
@@ -1,0 +1,1 @@
+flask
