+++ Project/static/styles/home.css
@@ -23,7 +23,8 @@ body{
 h1{
   color: RGB(34, 51, 59);
   text-align: center;
-
+  font-family: Georgia, 'Times New Roman', Times, serif;
+  font-size: 54px;
 }
 .center {
   display: block;
