+++ Project/templates/bookinfo.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        
+    </head>
+
+    <body>
+        <h1>{{book.b_title}}</h1>
+        <div>Total Checkouts: {{book.b_totalcheckouts}}</div>
+        <div>Total Holds: {{book.b_totalholds}}</div>
+    </body>
+</html>
