+++ Project/templates/home.html
@@ -7,7 +7,7 @@
 
 </head>
 
-<body>
+<body style="background-color:rgb(34, 51, 59)">
     <div class="box">
         <h1>
             Library for CS Nerds
