+++ Project/templates/librarian.html
@@ -21,7 +21,7 @@
     <h3>Books</h3>
     <input id="bookSearchfield" type="text">
     <button type="button" onclick="searchBooksByTitle()">Search</button>
-    
+    <div id="searchResCount"></div>
     <table>
         <thead>
             <tr class="align">
