--- page.html
@@ -1,19 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-    <head>
-
-        <title>Hello I'm your handy dandy database</title>
-        
-        
-        
-    </head>
-
-    <body>
-        <h1>
-            Hello I'm your handy dandy database system
-        </h1>
-    </body>
-
-
-</html>
