+++ setup.py
@@ -1,0 +1,10 @@
+from setuptools import setup
+
+setup(
+    name='Library Database Project',
+    packages=['Project'],
+    include_package_data=True,
+    install_requires=[
+        'flask',
+    ],
+)
