--- templates/bookinfo.html
@@ -1,10 +1,0 @@
-<!DOCTYPE html>
-<html>
-    <head>
-        
-    </head>
-
-    <body>
-        {{book.b_title}}
-    </body>
-</html>
